NAT traversal: add support for TCP relays
