+++ toxcore/network.c
@@ -45,17 +45,7 @@
 
 #include "network.h"
 #include "util.h"
-
-#ifdef HAVE_LIBMINIUPNPC
-#include <miniupnpc/miniupnpc.h>
-#include <miniupnpc/miniwget.h>
-#include <miniupnpc/upnpcommands.h>
-#include <miniupnpc/upnperrors.h>
-#endif
-
-#ifdef HAVE_LIBNATPMP
-#include <natpmp.h>
-#endif
+#include "nat_traversal.h"
 
 #if defined(_WIN32) || defined(__WIN32__) || defined (WIN32)
 
@@ -127,96 +117,6 @@ static int inet_pton(sa_family_t family, const char *addrString, void *addrbuf)
     return 0;
 }
 
-#endif
-
-#ifdef HAVE_LIBMINIUPNPC
-/* Setup port forwarding using UPnP */
-static void upnp_map_port(uint16_t port)
-{
-    LOGGER_DEBUG("Attempting to set up UPnP port forwarding");
-
-    int error = 0;
-    struct UPNPDev *devlist = NULL;
-
-#if MINIUPNPC_API_VERSION < 14
-    devlist = upnpDiscover(1000, NULL, NULL, 0, 0, &error);
-#else
-    devlist = upnpDiscover(1000, NULL, NULL, 0, 0, 2, &error);
-#endif
-
-    if (error) {
-        LOGGER_WARNING("UPnP discovery failed (error = %d)", error);
-        return;
-    }
-
-    struct UPNPUrls urls;
-    struct IGDdatas data;
-    char lanaddr[64];
-
-    error = UPNP_GetValidIGD(devlist, &urls, &data, lanaddr, sizeof(lanaddr));
-    freeUPNPDevlist(devlist);
-    if (error) {
-        if (error == 1) {
-            LOGGER_INFO("A valid IGD has been found.");
-
-            char portstr[10];
-            snprintf(portstr, sizeof(portstr), "%d", port);
-
-            error = UPNP_AddPortMapping(urls.controlURL, data.first.servicetype, portstr, portstr, lanaddr, "Tox", "UDP", 0, "0");
-            if (error) {
-                LOGGER_WARNING("UPnP port mapping failed (error = %d)", error);
-            } else {
-                LOGGER_INFO("UPnP mapped port %d", port);
-            }
-        } else if (error == 2) {
-            LOGGER_WARNING("IGD was found but reported as not connected.");
-        } else if (error == 3) {
-            LOGGER_WARNING("UPnP device was found but not recoginzed as IGD.");
-        } else {
-            LOGGER_WARNING("Unknown error finding IGD: %d", error);
-        }
-
-        FreeUPNPUrls(&urls);
-    } else {
-        LOGGER_WARNING("No IGD was found.");
-    }
-}
-#endif
-
-#ifdef HAVE_LIBNATPMP
-/* Setup port forwarding using NAT-PMP */
-static void natpmp_map_port(uint16_t port)
-{
-    LOGGER_DEBUG("Attempting to set up NAT-PMP port forwarding");
-
-    int error;
-    natpmp_t natpmp;
-    natpmpresp_t resp;
-
-    error = initnatpmp(&natpmp, 0, 0);
-    if (error) {
-        LOGGER_WARNING("NAT-PMP initialization failed (error = %d)", error);
-        return;
-    }
-
-    error = sendnewportmappingrequest(&natpmp, NATPMP_PROTOCOL_UDP, port, port, 3600);
-    if (error != 12) {
-        LOGGER_WARNING("NAT-PMP send request failed (error = %d)", error);
-        closenatpmp(&natpmp);
-        return;
-    }
-
-    error = readnatpmpresponseorretry(&natpmp, &resp);
-    for ( ; error == NATPMP_TRYAGAIN ; error = readnatpmpresponseorretry(&natpmp, &resp) )
-        sleep(1);
-
-    if (error)
-        LOGGER_WARNING("NAT-PMP port mapping failed (error = %d)", error);
-    else
-        LOGGER_INFO("NAT-PMP mapped port %d", port);
-
-    closenatpmp(&natpmp);
-}
 #endif
 
 /* Check if socket is valid.
@@ -818,12 +718,12 @@ Networking_Core *new_networking_nat(IP ip, uint16_t port_from, uint16_t port_to,
 
 #ifdef HAVE_LIBMINIUPNPC
             if ((traversal_type == TOX_TRAVERSAL_TYPE_UPNP) || (traversal_type == TOX_TRAVERSAL_TYPE_ALL))
-                upnp_map_port(ntohs(temp->port));
+                upnp_map_port(NAT_TRAVERSAL_UDP,ntohs(temp->port));
 #endif
 
 #ifdef HAVE_LIBNATPMP
             if ((traversal_type == TOX_TRAVERSAL_TYPE_NATPMP) || (traversal_type == TOX_TRAVERSAL_TYPE_ALL))
-                natpmp_map_port(ntohs(temp->port));
+                natpmp_map_port(NAT_TRAVERSAL_UDP,ntohs(temp->port));
 #endif
 
             return temp;
